performance: Refactor runner script
The script can now look through the compilers commit history to find
the point before and after a hook was merged. It can then sync `phobos`
to the same date as the commit prior to the introduction of the
template. This allows testing older hooks that have been already
integrated into the compiler.

Limitation: this still uses `dmd` which is bad at optimisations, but
more advanced backends (`gdc` and `ldc`) are slower to integrate hooks
and more difficult to control.

Signed-off-by: Teodor Dutu <teodor.dutu@gmail.com>